fw/boards/sf32lb52: use uint16_t for battery capacity

+++ src/fw/board/board_sf32lb52.h
@@ -155,7 +155,7 @@ typedef struct {
   //! Percentage for watch only mode
   const uint8_t low_power_threshold;
   //! Approximate hours of battery life
-  const uint8_t battery_capacity_hours;
+  const uint16_t battery_capacity_hours;
 } BoardConfigPower;
 
 typedef enum {
